requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.0
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.85.1
frozenlist==1.3.3
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
multidict==6.0.4
numpy==1.21.6
packaging==23.0
passlib==1.7.4
playwright==1.31.1
pluggy==1.0.0
psycopg2-binary==2.9.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyee==9.0.4
pymongo==4.3.3
pytest==7.2.2
pytest-base-url==2.0.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.2
redis==4.4.2
requests==2.28.1
rsa==4.9
scrapingbee==1.1.8
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
starlette==0.20.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
yarl==1.8.2
zipp==3.9.0
